First prototype of resolvers API, we are now fetching some first echonest data.
This data is now forwardet to one or many resolver, we need to implement some "query data"
API though so that we can query data from many locations (Echonest, Jamendo, Magnatune...)
so that we can get the full power of music.

The question is then how to decide what music we really need.  But this is something
we can discuss once we have more resolvers and are able to play more music on the web.